+++ trade-finance-manager-api/package.json
@@ -59,7 +59,6 @@
     "mongodb": "^4.17.2",
     "passport": "0.7.0",
     "passport-jwt": "^4.0.1",
-    "sanitize-html": "^2.17.5",
     "ts-node": "^10.9.2",
     "tslib": "^2.8.1",
     "validator": "^13.15.35",
